+++ server/packages/super_simple_authentication_postgres_data_storage/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1-dev.2
+
+- Make Postgres write statements directly (instead of using postgres_builder)
+- Add a setup function to write the schema directly to the database on startup
+
 ## 0.0.1-dev.1
 
 - Initial release!